refactor: plural over singular
fix: rename ui sub directory to containers
