+++ yatai/ui/src/components/BentoServiceDetail/ArtifactsTable.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as lodash from "lodash";
-import Table from "../../ui/Table";
-import { Section } from "../../ui/Layout";
+import Table from "../../containers/Table";
+import { Section } from "../../containers/Layout";
 
 interface IArtifactProps {
   name: string;
