+++ yatai/ui/src/components/DeploymentDetail/ConfigurationTable.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import * as lodash from "lodash";
 
-import Table from "../../ui/Table";
-import { Section } from "../../ui/Layout";
+import Table from "../../containers/Table";
+import { Section } from "../../containers/Layout";
 
 const parseKeyForDisplay = (key: string) => {
   return lodash.capitalize(key).replace("_", " ");
