+++ yatai/web_server/package.json
@@ -9,10 +9,10 @@
     "dev": "bash -c 'concurrently \"bentoml yatai-service-start --no-ui\"  \"nodemon -- ${0} ${1} ${2} ${3} ${4}\" \"npm run watch:client\"'",
     "watch": "npm:watch:*",
     "watch:server": "nodemon",
-    "watch:client": "cd client && yarn start",
+    "watch:client": "cd ../ui && yarn start",
     "build": "rm -rf dist && concurrently \"npm run build:client\" \"npm run build:server\"",
     "build:server": "webpack && cd dist && npm init --yes && yarn add express@\"~4.17.1\" protobufjs@\"~6.8.8\" && rm -rf node_modules/.bin",
-    "build:client": "cd client && yarn build && mkdir -p ../dist/client && cp -r build/* ../dist/client",
+    "build:client": "cd ../ui && yarn build && mkdir -p ../web_server/dist/client && cp -r build/* ../web_server/dist/client",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
